requirements.txt
flask
werkzeug
gunicorn
psycopg2-binary
python-dotenv
mercadopago
stripe
